+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Publish: add askBeforePublishing to supress asking for confirmation
+
 ## [3.0.5] - 2024-04-09
 
 ### Fixed
@@ -97,6 +103,7 @@
 
 - Initial version.
 
+[Unreleased]: https://github.com/inlavigo/gg_publish/compare/3.0.5...HEAD
 [3.0.5]: https://github.com/inlavigo/gg_publish/compare/3.0.4...3.0.5
 [3.0.4]: https://github.com/inlavigo/gg_publish/compare/3.0.3...3.0.4
 [3.0.3]: https://github.com/inlavigo/gg_publish/compare/3.0.2...3.0.3